requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
robotframework==7.4.1
robotframework-requests==0.9.7
urllib3==2.6.3
